feat: add new sections for business areas, contact, distinguish, hero, and values
