+++ src/components/ui/Checkmark.tsx
@@ -1,0 +1,17 @@
+export default function Checkmark({ className }: { className?: string }) {
+  return (
+    <svg
+      aria-hidden="true"
+      viewBox="0 0 24 24"
+      fill="none"
+      className={className}>
+      <path
+        d="M5 12.5 9.2 17 19 7"
+        stroke="currentColor"
+        strokeWidth="2"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+      />
+    </svg>
+  );
+}
